# fe/vdbas_exp_fe/apiContract/API-contract.yaml
openapi: 3.0.3
info:
  title: EXP CAPEX Dossier API
  description: |
    API quản lý **Hồ sơ Chi đầu tư (CAPEX)** – Phân hệ Chi (EXP) trong hệ thống VDBAS.

    ## Luồng trạng thái hồ sơ
    ```
    Start → DRAFT → SAVED → SUBMITTED → APPROVED (Checker) → APPROVED (Approver) → COMPLETED
                                ↘ REJECTED (Checker reject / Approver reject)
               ↘ CANCELLED (Maker xoá)
    ```

    ## Phân quyền
    - **Maker**: Tạo mới, Sửa, Xoá, Lưu nháp, Gửi kiểm soát (trong phạm vi TREASURY_CODE)
    - **Checker**: Kiểm soát (Approve/Reject) hồ sơ ở trạng thái SUBMITTED
    - **Approver**: Phê duyệt (Approve/Reject) hồ sơ ở trạng thái APPROVED (đã kiểm soát)
    - **Viewer**: Chỉ xem

    ## Lưu ý kỹ thuật
    - Mọi ID kiểu UUID (chuỗi 36 ký tự, dạng `xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx`)
    - Ngày dùng `yyyy-MM-dd`; DateTime dùng ISO 8601 `yyyy-MM-ddTHH:mm:ss+07:00`
    - Số tiền VND kiểu `integer` (đơn vị VND)
    - Optimistic lock qua trường `version` (gửi kèm khi Update)
    - Phòng double-submit: header `X-Idempotency-Key` (UUID do client sinh)
    - Phân quyền theo `TREASURY_CODE` lấy từ JWT claim
  version: "0.2.0"
  contact:
    name: VDBAS Development Team
servers:
  - url: /api/v1
    description: API version 1
security:
  - bearerAuth: []
...
paths:
...
  /exp/capex/dossiers:
    get:
      tags: [Dossiers]
      operationId: listDossiers
      summary: Danh sách hồ sơ (EXP.CAPEX_DOSSIER.LIST.SEARCH)
      description: |
        Tra cứu danh sách hồ sơ chi đầu tư theo nhiều tiêu chí.
        Kết quả giới hạn theo TREASURY_CODE của người dùng hiện tại.
        Mặc định sort theo CREATED_DATE DESC.
      parameters:
        - name: search
          in: query
          description: Tìm nhanh – tìm theo Người lập hoặc Tên dự án (P1-12)
          schema: { type: string, maxLength: 200 }
        - name: dossierCode
          in: query
          description: Mã hồ sơ (tìm chính xác hoặc bắt đầu bằng)
          schema: { type: string, maxLength: 100 }
        - name: projectCode
          in: query
          description: Mã dự án/công trình
          schema: { type: string, maxLength: 100 }
        - name: dateField
          in: query
          description: "Loại ngày lọc: SEND_DATE | CREATED_DATE | CHECKED_DATE | APPROVED_DATE"
          schema:
            type: string
            enum: [SEND_DATE, CREATED_DATE, CHECKED_DATE, APPROVED_DATE]
            default: SEND_DATE
        - name: fromDate
          in: query
          description: "Từ ngày (yyyy-MM-dd), áp dụng cho dateField đã chọn"
          schema: { type: string, format: date }
        - name: toDate
          in: query
          description: "Đến ngày (yyyy-MM-dd), phải ≥ fromDate"
          schema: { type: string, format: date }
        - name: fStatus
          in: query
          description: "Trạng thái hồ sơ – Multi-select, dùng nhiều lần (P1-13)"
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/DossierStatus'
        - name: dataSourceCode
          in: query
          description: "Nguồn gốc – Multi-select (P1-13)"
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, enum: [THU_CONG, DVC] }
        - name: createdBy
          in: query
          description: Username người lập
          schema: { type: string, maxLength: 100 }
        - name: page
          in: query
          description: Số trang (bắt đầu từ 1)
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          description: Số bản ghi/trang
          schema:
            type: integer
            enum: [20, 50, 100, 200]
            default: 20
        - name: sortBy
          in: query
          description: Trường sắp xếp
          schema:
            type: string
            enum: [DOSSIER_CODE, PROJECT_CODE, SEND_DATE, CREATED_DATE, F_STATUS]
            default: CREATED_DATE
        - name: sortDir
          in: query
          description: Chiều sắp xếp
          schema: { type: string, enum: [ASC, DESC], default: DESC }
      responses:
        '200':
          description: Danh sách hồ sơ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Dossiers]
      operationId: createDossier
      summary: Tạo mới & Lưu hồ sơ → SAVED (EXP.CAPEX_DOSSIER.NEW.SAVE)
      description: |
        Validate cơ bản; sinh DOSSIER_CODE; lưu hồ sơ ở trạng thái `SAVED`.
        Ghi audit log. Trả về MSG-OK-SAVE (VDBAS-EXP-SUC01).
        Yêu cầu vai trò Maker.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DossierCreateRequest'
      responses:
        '201':
          description: Tạo hồ sơ thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'

  # ═══════════════════════════════════════════
  #  DOSSIER – LƯU NHÁP
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/drafts:
    post:
      tags: [Dossiers]
      operationId: saveDossierDraft
      summary: Lưu nháp → DRAFT (EXP.CAPEX_DOSSIER.NEW.SAVE_DRAFT)
      description: |
        Validate tối thiểu (chỉ kiểm tra định dạng, không bắt buộc đủ trường).
        Lưu hồ sơ ở trạng thái `DRAFT`. Ghi audit log.
        Yêu cầu vai trò Maker.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DossierDraftRequest'
      responses:
        '201':
          description: Lưu nháp thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ═══════════════════════════════════════════
  #  DOSSIER – CHI TIẾT / SỬA / XOÁ
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/{dossierId}:
    parameters:
      - $ref: '#/components/parameters/DossierId'

    get:
      tags: [Dossiers]
      operationId: getDossier
      summary: Xem chi tiết hồ sơ (EXP.CAPEX_DOSSIER.VIEW.OPEN)
      description: |
        Trả về đầy đủ thông tin hồ sơ: thông tin chung (§B1.1), danh sách chứng từ (§B1.2),
        đính kèm, lịch sử giao dịch, trạng thái phê duyệt.
        Tất cả vai trò có quyền theo TREASURY_CODE.
      responses:
        '200':
          description: Chi tiết hồ sơ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Dossiers]
      operationId: updateDossier
      summary: Cập nhật hồ sơ (EXP.CAPEX_DOSSIER.EDIT.SAVE)
      description: |
        Chỉ áp dụng khi F_STATUS ∈ {DRAFT, SAVED} và NSD là Maker gốc.
        Optimistic lock: `version` phải khớp với DB (VAL-15); F-VER tăng +1 sau khi lưu.
        Ghi audit oldValue→newValue. Trạng thái không đổi.
        Yêu cầu vai trò Maker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DossierUpdateRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Optimistic lock conflict (VAL-15 / VDBAS-EXP-0005)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errorCode: VDBAS-EXP-0005
                message: "Bản ghi đã bị thay đổi từ phiên khác. Vui lòng tải lại trước khi tiếp tục"
                traceId: "abc-123"

    delete:
      tags: [Dossiers]
      operationId: deleteDossier
      summary: Xoá mềm hồ sơ → CANCELLED (EXP.CAPEX_DOSSIER.DELETE.CONFIRM)
      description: |
        Soft-delete: STATUS=0, F_STATUS=CANCELLED. Ghi audit. Trả về MSG-OK-DELETE.
        Chỉ áp dụng khi F_STATUS ∈ {DRAFT, SAVED} và NSD là Maker gốc.
        Yêu cầu vai trò Maker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteDossierRequest'
      responses:
        '200':
          description: Xoá thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                errorCode: VDBAS-EXP-SUC03
                message: "Xoá hồ sơ thành công"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Trạng thái không cho phép xoá (VAL-13 / VDBAS-EXP-0003)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════
  #  WORKFLOW – GỬI KIỂM SOÁT
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/{dossierId}/submit:
    post:
      tags: [Workflow]
      operationId: submitDossier
      summary: Gửi kiểm soát SAVED → SUBMITTED (EXP.CAPEX_DOSSIER.NEW.SUBMIT)
      description: |
        Validate đầy đủ (§A6); chuyển trạng thái SAVED/VALIDATED → SUBMITTED;
        sinh HASH_INFO; assign Checker; gửi notification Checker; ghi audit.
        Điều kiện: hồ sơ có ít nhất 1 chứng từ hợp lệ (BIZ-012).
        Phòng double-submit qua X-Idempotency-Key.
        Yêu cầu vai trò Maker.
      parameters:
        - $ref: '#/components/parameters/DossierId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: integer
                  description: Phiên bản optimistic lock hiện tại của hồ sơ
              required: [version]
      responses:
        '200':
          description: Gửi kiểm soát thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowActionResponse'
              example:
                success: true
                errorCode: VDBAS-EXP-SUC02
                message: "Đã gửi hồ sơ để kiểm soát/phê duyệt"
                data:
                  dossierId: "550e8400-e29b-41d4-a716-446655440000"
                  fStatus: SUBMITTED
                  assignUser: "checker_user01"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: |
            Không đáp ứng điều kiện gửi kiểm soát:
            - Chưa có chứng từ (BIZ-012)
            - Trạng thái không hợp lệ (VAL-13 / VDBAS-EXP-0003)
            - Không dành được dự toán (VAL-21 / VDBAS-EXP-0016)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════
  #  WORKFLOW – PHÊ DUYỆT / KIỂM SOÁT
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/{dossierId}/approve:
    post:
      tags: [Workflow]
      operationId: approveDossier
      summary: Kiểm soát / Phê duyệt hồ sơ (EXP.CAPEX_DOSSIER.APPROVE.APPROVE)
      description: |
        - **Checker** (F_STATUS=SUBMITTED): chuyển → APPROVED "Đã kiểm soát"; assign Approver; notify Approver.
        - **Approver** (F_STATUS=APPROVED đã kiểm soát): chuyển → APPROVED "Đã phê duyệt"; trigger luồng kế tiếp; notify Maker.
        SoD bắt buộc: mỗi cấp khác user và khác vai trò (BIZ-001).
        Ghi EXP_APPROVAL_LOG, EXP_DIGITAL_SIGNED (nếu có ký số).
      parameters:
        - $ref: '#/components/parameters/DossierId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        '200':
          description: Phê duyệt thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ═══════════════════════════════════════════
  #  WORKFLOW – TỪ CHỐI
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/{dossierId}/reject:
    post:
      tags: [Workflow]
      operationId: rejectDossier
      summary: Từ chối kiểm soát / phê duyệt (EXP.CAPEX_DOSSIER.APPROVE.REJECT)
      description: |
        - **Checker** từ chối: SUBMITTED → REJECTED; assign Maker; notify Maker.
        - **Approver** từ chối: APPROVED (đã kiểm soát) → REJECTED; assign Checker; notify Checker.
        Lý do từ chối bắt buộc ≥ 10 ký tự (BIZ-006). Ghi EXP_APPROVAL_LOG.
      parameters:
        - $ref: '#/components/parameters/DossierId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
      responses:
        '200':
          description: Từ chối thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ═══════════════════════════════════════════
  #  DOSSIER – SAO CHÉP
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/{dossierId}/copy:
    post:
      tags: [Dossiers]
      operationId: copyDossier
      summary: Sao chép hồ sơ thành bản nháp mới (EXP.CAPEX_DOSSIER.NEW.COPY)
      description: |
        Tạo bản sao của hồ sơ (thông tin chung + danh sách chứng từ) ở trạng thái DRAFT.
        Sinh DOSSIER_CODE mới; CREATED_BY = user hiện tại; CREATED_DATE = now.
        Yêu cầu vai trò Maker.
      parameters:
        - $ref: '#/components/parameters/DossierId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Sao chép thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierCreateResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════
  #  DOCUMENTS – CHỨNG TỪ TRONG HỒ SƠ
  # ═══════════════════════════════════════════
...
  /exp/capex/dossiers/{dossierId}/approval-log:
    parameters:
      - $ref: '#/components/parameters/DossierId'
    get:
      tags: [Audit]
      operationId: getDossierApprovalLog
      summary: Lịch sử phê duyệt (EXP.CAPEX_DOSSIER.VIEW.APPROVAL)
      description: Trả về luồng Maker→Checker→Approver với trạng thái từng bước.
      responses:
        '200':
          description: Lịch sử phê duyệt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalLogResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════
  #  AUDIT – LỊCH SỬ THAY ĐỔI
  # ═══════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT chứa claims userId, role, treasuryCode
  parameters:
    DossierId:
      name: dossierId
      in: path
      required: true
      description: UUID hồ sơ (EXP_DOSSIER.ID)
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      description: UUID do client sinh – phòng chống double-submit
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            errorCode: VDBAS-VAL-0002
            message: "Dữ liệu đầu vào không hợp lệ"
            traceId: "xyz-456"
            fieldErrors:
              - field: projectCode
                errorCode: VDBAS-VAL-0002
                message: "Vui lòng nhập Mã dự án/công trình"
    Unauthorized:
      description: Chưa xác thực (VDBAS-AUT-0001)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errorCode: VDBAS-AUT-0002
            message: "Phiên đăng nhập đã hết hạn, vui lòng đăng nhập lại"
            traceId: "xyz-789"
    Forbidden:
      description: Không có quyền thực hiện (VDBAS-AUT-0001)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errorCode: VDBAS-AUT-0001
            message: "Bạn không có quyền thực hiện thao tác này"
            traceId: "abc-789"
    NotFound:
      description: Không tìm thấy bản ghi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Dữ liệu trùng lặp (VDBAS-EXP-0002)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errorCode: VDBAS-EXP-0002
            message: "Đã tồn tại bản ghi có DOSSIER_CODE = 'EXP/CAPEX/2025/00001'"
            traceId: "abc-001"
    UnprocessableEntity:
      description: Vi phạm quy tắc nghiệp vụ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DossierStatus:
      type: string
      description: |
        Trạng thái hồ sơ (enum chuẩn §A11.1):
        - DRAFT: Lưu nháp (badge xám)
        - SAVED: Đã lưu (badge xám)
        - VALIDATED: Đã kiểm tra — nhận tự động (badge xanh dương)
        - SUBMITTED: Đã gửi phê duyệt (badge xanh dương)
        - APPROVED: Đã phê duyệt / Đã kiểm soát (badge xanh lá)
        - REJECTED: Đã từ chối (badge đỏ)
        - COMPLETED: Đã hoàn thành (badge xanh lá)
        - CANCELLED: Đã huỷ (badge cam)
      enum:
        - DRAFT
        - SAVED
        - VALIDATED
        - SUBMITTED
        - APPROVED
        - REJECTED
        - COMPLETED
        - CANCELLED
...
    DossierCreateRequest:
      type: object
      required:
        - sendDate
        - dataSourceCode
        - projectCode
        - investorCode
        - projectManagementCode
      properties:
        sendDate:
          type: string
          format: date
          description: "Ngày gửi hồ sơ (yyyy-MM-dd); mặc định ngày hiện tại"
          example: "2025-06-17"
        dataSourceCode:
          type: string
          enum: [THU_CONG, DVC]
          default: THU_CONG
          description: "Nguồn gốc hồ sơ: THU_CONG=Thủ công (mặc định), DVC=Dịch vụ công"
        projectCode:
          type: string
          maxLength: 100
          description: "Mã dự án/công trình (LOV.01)"
          example: "7122155"
        projectSpecificCode:
          type: string
          maxLength: 100
          nullable: true
          description: "Mã dự án đặc thù; chỉ áp dụng khi Project_Type=MILITARY"
          example: "001200037"
        investorCode:
          type: string
          maxLength: 100
          description: "Mã chủ đầu tư (LOV.09 / LOV.13); auto-fill từ project"
          example: "1059227"
        projectManagementCode:
          type: string
          maxLength: 100
          description: "Mã ban QLDA / ĐVQHNS (LOV.05)"
          example: "1059227"
    DossierDraftRequest:
      type: object
      description: Lưu nháp — chỉ validate định dạng, không bắt buộc đủ trường bắt buộc
      properties:
        sendDate:
          type: string
          format: date
          example: "2025-06-17"
        dataSourceCode:
          type: string
          enum: [THU_CONG, DVC]
          default: THU_CONG
        projectCode:
          type: string
          maxLength: 100
        projectSpecificCode:
          type: string
          maxLength: 100
          nullable: true
        investorCode:
          type: string
          maxLength: 100
        projectManagementCode:
          type: string
          maxLength: 100
    DossierUpdateRequest:
      type: object
      required: [version]
      description: |
        Cập nhật hồ sơ. Trường `version` bắt buộc để kiểm tra optimistic lock.
        Các trường immutable (DOSSIER_CODE, CREATED_BY, CREATED_DATE) không được gửi — backend sẽ reject.
      properties:
        version:
          type: integer
          description: "Phiên bản F-VER hiện tại (optimistic lock — VAL-15)"
          example: 1
        sendDate:
          type: string
          format: date
        projectCode:
          type: string
          maxLength: 100
        projectSpecificCode:
          type: string
          maxLength: 100
          nullable: true
        investorCode:
          type: string
          maxLength: 100
        projectManagementCode:
          type: string
          maxLength: 100
    DeleteDossierRequest:
      type: object
      required:
        - deleteReason
        - confirmReviewed
      properties:
        deleteReason:
          type: string
          minLength: 10
          maxLength: 500
          description: "Lý do xoá — bắt buộc ≥ 10 ký tự (VAL-16)"
          example: "Hồ sơ nhập sai thông tin dự án, cần tạo lại từ đầu"
        confirmReviewed:
          type: boolean
          description: "Xác nhận đã rà soát trước khi xoá — bắt buộc = true (VAL-16)"
          example: true
    ApproveRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          description: Ghi chú phê duyệt (không bắt buộc)
        digitalSign:
          $ref: '#/components/schemas/DigitalSignInfo'
    RejectRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 10
          maxLength: 500
          description: "Lý do từ chối — bắt buộc ≥ 10 ký tự (BIZ-006)"
          example: "Chứng từ chưa đầy đủ, thiếu hóa đơn gốc số HĐ-2025-0042"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        errorCode:
          type: string
          description: Mã thông báo thành công (VDBAS-EXP-SUCxx)
          example: VDBAS-EXP-SUC01
        message:
          type: string
          example: "Lưu hồ sơ thành công"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errorCode:
          type: string
          description: "Mã lỗi chuẩn (VDBAS-VAL-*, VDBAS-EXP-*, VDBAS-SYS-*, VDBAS-AUT-*)"
          example: VDBAS-EXP-0003
        message:
          type: string
          example: "Hồ sơ đang ở trạng thái SUBMITTED, không cho phép Sửa/Xoá"
        traceId:
          type: string
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalRecords:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
...
    DossierListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/DossierSummary'
            pagination:
              $ref: '#/components/schemas/Pagination'
            totalBaseAmount:
              type: integer
              description: "Tổng tiền VND tất cả hồ sơ khớp bộ lọc (status ≠ CANCELLED)"
            statusCounts:
              type: array
              items:
                $ref: '#/components/schemas/StatusCount'
              description: Thống kê số bản ghi theo trạng thái (§B2.3 P2-11)
    DossierCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        errorCode:
          type: string
          example: VDBAS-EXP-SUC01
        message:
          type: string
          example: "Lưu hồ sơ thành công"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            dossierCode:
              type: string
              example: "EXP/CAPEX/2025/00001"
            fStatus:
              $ref: '#/components/schemas/DossierStatus'
            version:
              type: integer
              example: 1
    DossierUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        errorCode:
          type: string
          example: VDBAS-EXP-SUC01
        message:
          type: string
          example: "Lưu hồ sơ thành công"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            dossierCode:
              type: string
            fStatus:
              $ref: '#/components/schemas/DossierStatus'
            version:
              type: integer
              description: F-VER mới sau khi cập nhật (+1)
    DossierDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/DossierDetail'
    WorkflowActionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        errorCode:
          type: string
          example: VDBAS-EXP-SUC02
        message:
          type: string
          example: "Đã gửi hồ sơ để kiểm soát/phê duyệt"
        data:
          type: object
          properties:
            dossierId:
              type: string
              format: uuid
            fStatus:
              $ref: '#/components/schemas/DossierStatus'
            fStatusName:
              type: string
            assignUser:
              type: string
              description: User được phân công xử lý tiếp theo
...
    ApprovalLogResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            dossierId:
              type: string
              format: uuid
            currentStatus:
              $ref: '#/components/schemas/DossierStatus'
            workflow:
              type: array
              items:
                $ref: '#/components/schemas/ApprovalLogEntry'
...